+++ src/games/RiichiMahjong.tsx
@@ -581,8 +581,8 @@ export const RiichiMahjong: React.FC = () => {
         gameId="riichi-mahjong"
         gameName="Riichi Mahjong (2-Player Duel)"
         turn={turn}
-        statusText={`Player ${turn}'s Turn`}
-        subStatusText={statusMessage}
+        onRestart={resetGame}
+        statusMessage={statusMessage}
       />
 
       {/* Main Mat Arena - Responsive scaling */}